Added Recurrent complex to dep + convert 19 dungeons to .rcst file
Optimize imports
Remove some redundant casts

+++ src/main/java/com/gibby/dungeon/blocks/TileEntityRedKeyOpen.java
@@ -2,8 +2,8 @@
 
 package com.gibby.dungeon.blocks;
 
-import net.minecraft.tileentity.*;
-import com.gibby.dungeon.*;
+import com.gibby.dungeon.Dungeons;
+import net.minecraft.tileentity.TileEntity;
 
 public class TileEntityRedKeyOpen extends TileEntity
 {
@@ -27,7 +27,7 @@ public class TileEntityRedKeyOpen extends TileEntity
             for (int Z = -1; Z < 2; ++Z) {
                 for (int Y = -1; Y < 3; ++Y) {
                     this.worldObj.setBlockToAir(x, y + Y, z + Z);
-                    this.worldObj.playSound((double)x, (double)y, (double)z, "mob.irongolem.hit", 1.0f, 1.0f, false);
+                    this.worldObj.playSound(x, y, z, "mob.irongolem.hit", 1.0f, 1.0f, false);
                     for (int i = 0; i < 10; ++i) {
                         this.worldObj.spawnParticle("reddust", x + Dungeons.randGauss(), y + Y + Dungeons.randGauss(), z + Z + Dungeons.randGauss(), 0.0, 0.0, 0.0);
                     }
